Move LogsService implementations from .utils to .logs.
